Added Geocoding API, code modularization and improvements
