+++ index.html
@@ -32,18 +32,6 @@
             <div class="footer_btn" id="ftr_btn_3" onclick="window.location='/views/googleFit.html'">Playground</div>
       </div>
       <script src="lib/jquery-3.3.1.min.js"></script>
-      <script src="lib/bootstrap.min.js"></script>
-      <script src="scripts/apiKeys.js"></script>
-      <script src="services/maps.js"></script>
-      <script src="scripts/fit.js"></script>
-      <script type="module" src="scripts/countryNames.js"></script>
-      <script type="module" src="modules/weather.js"></script>
-      <script type="module" src="modules/appendIntoTable.js"></script>
-      <script type="module" src="modules/weatherService.js"></script>
-      <script type="module" src="modules/locationService.js"></script>
-      <script type="module" src="scripts/fitHandleClick.js"></script>
-      <script type="module" src="modules/GoogleFit/fetchFitData.js"></script>
-      <script type="module" src="app.js"></script>
       <script>
           $('.footer_btn').removeClass('active');
           $('#ftr_btn_1').addClass('active');
