+++ styles.css
@@ -86,7 +86,7 @@ body {
     border-width: 3px;
 }
 
-#tabularData, #spinner, .fitnessDataContainer {
+#weatherTable, #spinner, .fitnessDataContainer {
     display: none;
 }
 
@@ -151,8 +151,7 @@ tr {
     width:300px;
     height:500px;
     background:#383826;
-    margin-bottom: 60px;
-    display: none;
+    margin-bottom: 15px;
 }
 
 .dataCards {
